feat: add actions on all participants header #20
