+++ web/src/pages/Participants.svelte
@@ -1,23 +1,16 @@
 <script>
-  import dayjs from "dayjs";
-  import { mutate } from "svelte-apollo";
   import { query } from "svelte-apollo";
-  import Input from "../components/base/Input.svelte";
-  import Label from "../components/base/Label.svelte";
-  import Modal from "../components/overlays/Modal.svelte";
-  import { notify } from "../components/overlays/Notification.svelte";
+
   import ParticipantList from "../components/participants/ParticipantList.svelte";
+  import Import from "../components/participants/Import.svelte";
   import Layout from "../layouts/Layout.svelte";
   import { client } from "../lib/apollo";
-  import { participantPerQueryType } from "../lib/models/participants/participants.js";
   import {
-    getParticipants,
-    participantsExportFormat,
-    setValue,
+    participantPerQueryType,
+    exportCSV,
+    exportJson,
   } from "../lib/models/participants/participants.js";
-  import { ADD_PARTICIPANTS, GET_PROJECT_PARTICIPANTS } from "../lib/queries";
-  import { toCSV } from "../utils/csv.js";
-  import { download } from "../utils/download.js";
+  import { GET_PROJECT_PARTICIPANTS } from "../lib/queries";
   import { handleErrorMessage } from "../utils/errorQuery.js";
 
   const queryArgs = (project) => ({
@@ -28,98 +21,8 @@
   let participants;
   let allParticipants;
   let keys;
-  let files = [];
-  let isImportOpen = false;
-  let customKey;
-  let customValue;
   let projectID;
-
-  function handleCancelImport() {
-    isImportOpen = false;
-    files = [];
-    customKey = null;
-    customValue = null;
-  }
-
-  function handleImport() {
-    let file = files.length > 0 ? files[0] : null;
-    let isCustomEmpty;
-    let customData = {};
-
-    if (!file) {
-      notify({
-        failed: true,
-        title: `Could not import participants.`,
-        body: "No file selected.",
-      });
-      return;
-    }
-
-    if (customKey || customValue) {
-      if (
-        !customKey ||
-        !customKey.trim() ||
-        !customValue ||
-        !customValue.trim()
-      ) {
-        isCustomEmpty = true;
-      }
-
-      if (isCustomEmpty) {
-        notify({
-          failed: true,
-          title: `Could not import participants.`,
-          body: "Custom key/value pair can't be empty.",
-        });
-        return;
-      }
-
-      customData = { [customKey]: setValue(customValue) };
-    }
-
-    getParticipants(file, customData, async (newParticipants, error) => {
-      if (error) {
-        notify({
-          failed: true,
-          title: `Could not import participants.`,
-          body: error,
-        });
-        return;
-      }
-
-      try {
-        await mutate(client, {
-          mutation: ADD_PARTICIPANTS,
-          variables: {
-            input: {
-              participants: newParticipants,
-              projectID,
-            },
-          },
-        });
-
-        files = [];
-        isImportOpen = false;
-        customKey = null;
-        customValue = null;
-        notify({
-          failed: false,
-          title: `Participants imported.`,
-        });
-        setTimeout(() => {
-          location.reload();
-        }, 2000);
-      } catch (error) {
-        handleErrorMessage(error);
-        notify({
-          failed: true,
-          title: `Could not import participants`,
-          body:
-            "Something happened on the server, and we could not import the participants.",
-        });
-      }
-    });
-  }
+  let isOpen = false;
 
   async function fetchAllParticipants(project) {
     let args = queryArgs(project);
@@ -142,26 +45,16 @@
     switch (action) {
       case "exportjson": {
         await fetchAllParticipants(project);
-        const out = participantsExportFormat(allParticipants, keys);
-        const content = JSON.stringify(out);
-        const mime = "application/json;charset=utf-8";
-        const date = dayjs().format("YYYY-MM-DDTHH:mm:ss");
-        const filename = `Empirica recruitment export – ${date}.json`;
-        download(content, filename, mime);
+        exportJson(allParticipants, keys);
         break;
       }
       case "exportcsv": {
         await fetchAllParticipants(project);
-        const out = participantsExportFormat(allParticipants, keys, true);
-        const content = toCSV(out);
-        const mime = "text/csv;charset=utf-8";
-        const date = dayjs().format("YYYY-MM-DDTHH:mm:ss");
-        const filename = `Empirica recruitment export – ${date}.csv`;
-        download(content, filename, mime);
+        exportCSV(allParticipants, keys);
         break;
       }
       case "import":
-        isImportOpen = true;
+        isOpen = true;
         projectID = project.id;
         break;
       default:
@@ -204,27 +97,5 @@
     queryArgs={queryArgs(project)}
     bind:participants
     bind:keys />
-  <Modal
-    title="Import Participants"
-    button="Import"
-    bind:open={isImportOpen}
-    handleAccept={handleImport}
-    handleCancel={handleCancelImport}>
-    <div class="sm:flex flex-row sm:items-start">
-      <input type="file" bind:files />
-    </div>
-    <div class="mt-4">
-      <Label
-        text="Custom Data"
-        question="Custom data (key value pair) that will be created on imported partcipants." />
-    </div>
-    <div class="sm:flex flex-row sm:items-start">
-      <div class="mt-1">
-        <Input bind:value={customKey} placeholder="key" />
-      </div>
-      <div class="mt-1 ml-3">
-        <Input bind:value={customValue} placeholder="value" />
-      </div>
-    </div>
-  </Modal>
+  <Import {projectID} bind:isOpen />
 </Layout>
